backend/requirements.txt
Flask
Flask-RESTful
Flask-SQLAlchemy
Flask-Cors
python-dotenv
